+++ apps/heureka/src/components/common/SeverityCount.tsx
@@ -11,6 +11,7 @@ import {
   TooltipContent,
   KnownIcons,
   BadgeVariantType,
+  Icon,
 } from "@cloudoperators/juno-ui-components"
 
 type SeverityCountProps = {