+++ packages/ui-components/src/components/Icon/Icon.stories.ts
@@ -379,6 +379,7 @@ export const SeverityLow: Story = {
   args: {
     ...Default.args,
     icon: "severityLow",
+    color: "jn:text-theme-severity-low",
   },
 }
 
@@ -386,6 +387,7 @@ export const SeverityMedium: Story = {
   args: {
     ...Default.args,
     icon: "severityMedium",
+    color: "jn:text-theme-severity-medium",
   },
 }
 
@@ -393,6 +395,15 @@ export const SeverityHigh: Story = {
   args: {
     ...Default.args,
     icon: "severityHigh",
+    color: "jn:text-theme-severity-high",
+  },
+}
+
+export const SeverityVeryHigh: Story = {
+  args: {
+    ...Default.args,
+    icon: "severityVeryHigh",
+    color: "jn:text-theme-severity-very-high",
   },
 }
 
@@ -400,6 +411,15 @@ export const SeverityCritical: Story = {
   args: {
     ...Default.args,
     icon: "severityCritical",
+    color: "jn:text-theme-severity-critical",
+  },
+}
+
+export const SeverityUnknown: Story = {
+  args: {
+    ...Default.args,
+    icon: "severityUnknown",
+    color: "jn:text-theme-severity-unknown",
   },
 }
 
@@ -407,6 +427,34 @@ export const Success: Story = {
   args: {
     ...Default.args,
     icon: "success",
+  },
+}
+
+export const SortShortWideArrowUp: Story = {
+  args: {
+    ...Default.args,
+    icon: "sortShortWideArrowUp",
+  },
+}
+
+export const SortShortWideArrowDown: Story = {
+  args: {
+    ...Default.args,
+    icon: "sortShortWideArrowDown",
+  },
+}
+
+export const SortWideShortArrowUp: Story = {
+  args: {
+    ...Default.args,
+    icon: "sortWideShortArrowUp",
+  },
+}
+
+export const SortWideShortArrowDown: Story = {
+  args: {
+    ...Default.args,
+    icon: "sortWideShortArrowDown",
   },
 }